+++ griptape/config/openai_structure_config.py
@@ -1,3 +1,5 @@
+import os
+from typing import Optional
 from attrs import Factory, define, field
 
 from griptape.config import StructureConfig
@@ -17,25 +19,68 @@ from griptape.drivers import (
 
 @define
 class OpenAiStructureConfig(StructureConfig):
+    api_key: Optional[str] = field(kw_only=True, default=Factory(lambda: os.getenv("OPENAI_API_KEY")))
+    base_url: Optional[str] = field(default=None, kw_only=True, metadata={"serializable": True})
+    organization: Optional[str] = field(default=None, kw_only=True, metadata={"serializable": True})
+
     prompt_driver: BasePromptDriver = field(
-        default=Factory(lambda: OpenAiChatPromptDriver(model="gpt-4o")), metadata={"serializable": True}, kw_only=True
+        default=Factory(
+            lambda self: OpenAiChatPromptDriver(
+                model="gpt-4o", api_key=self.api_key, base_url=self.base_url, organization=self.organization
+            ),
+            takes_self=True,
+        ),
+        metadata={"serializable": True},
+        kw_only=True,
     )
     image_generation_driver: BaseImageGenerationDriver = field(
-        default=Factory(lambda: OpenAiImageGenerationDriver(model="dall-e-2", image_size="512x512")),
+        default=Factory(
+            lambda self: OpenAiImageGenerationDriver(
+                model="dall-e-2",
+                image_size="512x512",
+                api_key=self.api_key,
+                base_url=self.base_url,
+                organization=self.organization,
+            ),
+            takes_self=True,
+        ),
         kw_only=True,
         metadata={"serializable": True},
     )
     image_query_driver: BaseImageQueryDriver = field(
-        default=Factory(lambda: OpenAiImageQueryDriver(model="gpt-4o")), kw_only=True, metadata={"serializable": True}
+        default=Factory(
+            lambda self: OpenAiImageQueryDriver(
+                model="gpt-4o", api_key=self.api_key, base_url=self.base_url, organization=self.organization
+            ),
+            takes_self=True,
+        ),
+        kw_only=True,
+        metadata={"serializable": True},
     )
     embedding_driver: BaseEmbeddingDriver = field(
-        default=Factory(lambda: OpenAiEmbeddingDriver(model="text-embedding-3-small")),
+        default=Factory(
+            lambda self: OpenAiEmbeddingDriver(
+                model="text-embedding-3-small",
+                api_key=self.api_key,
+                base_url=self.base_url,
+                organization=self.organization,
+            ),
+            takes_self=True,
+        ),
         metadata={"serializable": True},
         kw_only=True,
     )
     vector_store_driver: BaseVectorStoreDriver = field(
         default=Factory(
-            lambda: LocalVectorStoreDriver(embedding_driver=OpenAiEmbeddingDriver(model="text-embedding-3-small"))
+            lambda self: LocalVectorStoreDriver(
+                embedding_driver=OpenAiEmbeddingDriver(
+                    model="text-embedding-3-small",
+                    api_key=self.api_key,
+                    base_url=self.base_url,
+                    organization=self.organization,
+                )
+            ),
+            takes_self=True,
         ),
         kw_only=True,
         metadata={"serializable": True},